backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
python-jose==3.3.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.97.0
python-jose[cryptography]==3.3.0 